fixed AG email bug and and added AG ministry check button, fixed read only admin role bugs
