+++ src/components/forms/MinistryInput.js
@@ -1,38 +1,20 @@
-import React, { useState, useEffect } from "react";
+import React, { useState } from "react";
 import InputLabel from "@mui/material/InputLabel";
 import MenuItem from "@mui/material/MenuItem";
 import FormControl from "@mui/material/FormControl";
 import Select from "@mui/material/Select";
-import { ministriesNames, checkBoxMinistries } from "../common/Constants";
+import { ministriesNames} from "../common/Constants";
 import TitleTypography from "../common/TitleTypography";
 import FormHelperText from "@mui/material/FormHelperText";
 import RequiredField from "../common/RequiredField";
 import Box from "@mui/material/Box";
 import Typography from "@mui/material/Typography";
-import Checkbox from '@mui/material/Checkbox';
-import FormControlLabel from '@mui/material/FormControlLabel';
 
+const MinistryInput = ({ formik,isDisabled, setAGministries }) => {
 
-const MinistryInput = ({ formik, isDisabled }) => {
-const MinistryInput = ({ formik, setAGministries }) => {
-  const [AGministry, setAGministry] = useState(false)
-
-  useEffect(() => {
-    setAGministries(checkBoxMinistries.indexOf(formik.values.ministry) !== -1?!AGministry:false)
-    
-  })
-
-  useEffect(() => {
-    setAGministries(checkBoxMinistries.indexOf(formik.values.ministry) !== -1 ? !AGministry : false)
- 
-  }, [AGministry, formik.values.ministry, setAGministries])
-  
-  const handleChangeCheckBox = (event) => {
-    setAGministry(event.target.checked)
-  }
  
   return (
-    <Box sx={{ mb: 5, mt: 5, mr: 10, width: 300 }}>
+    <Box sx={{ mb: 1, mt: 5, mr: 10, width: 300 }}>
       <div>
         <TitleTypography>Ministry</TitleTypography>
         <Typography sx={{ mb: 2 }} color="text.primary">
@@ -57,7 +39,7 @@ const MinistryInput = ({ formik, setAGministries }) => {
           name="ministry"
           label="Ministry"
           labelId="select-ministry"
-          disabled={false}
+          disabled={isDisabled}
           value={formik.values.ministry}
           onChange={formik.handleChange}
           error={formik.touched.ministry && Boolean(formik.errors.ministry)}
@@ -70,38 +52,8 @@ const MinistryInput = ({ formik, setAGministries }) => {
         </Select>
         <FormHelperText>
           {formik.touched.ministry && <RequiredField />}
-        </FormHelperText>
-        {
-          checkBoxMinistries.indexOf(formik.values.ministry) !== -1 &&
-          <FormControlLabel
-            style={{
-              minWidth: '700px',
-            }}
-            control={
-              <Checkbox
-                id="ministryAG"
-                name="ministryAGname"
-                type="checkbox"
-                checked={checkBoxMinistries.indexOf(formik.values.ministry) !== -1 ? AGministry : true}
-                onChange={handleChangeCheckBox}
-                style={{
-                  color: `${!AGministry?'red':'#003366'}`,
-                }}
-              />
-            }
-            label={
-              <Typography sx={{ width: 700, mt: 1 }} variant="subtitle2" color="text.secondary">
-                I confirm that I have contacted the AG Security and received their approval for provisioning the namespaces in Private Cloud Openshift plaform.
-              </Typography>
-            }
-          />}
-      </FormControl>
-      <Typography sx={{ width: 700, mt: 1 }} variant="subtitle2" color="text.secondary">
-        * All product teams from the Ministries of Attorney General, Public Safety
-        and Solicitor General and Emergency Management BC and BC Housing must
-        engage with <a target="_blank" rel="noopener noreferrer" href="mailto: JAGMISO@gov.bc.ca">AG Security</a> to
-prior to submitting a request for a new product.
-      </Typography>
+        </FormHelperText>       
+      </FormControl>   
     </Box>
   );
 };
